Moved Autocomplete to CLI
