+++ src/Views/cli_view.py
@@ -1,13 +1,10 @@
-from prompt_toolkit.completion import NestedCompleter
 from Models.uml_diagram import UML_Diagram
 from Models.uml_list import UML_List_Visitor
-from Views.relation_completer import RelationCompleter
 
 class CLI_View:
     
     def __init__(self):
         self._lister = UML_List_Visitor() 
-        self._completer = self.setup_autocomplete() 
              
     #A list of surface level commands    
     def help(self):
@@ -77,52 +74,3 @@ class CLI_View:
         return ''.join(r.accept(self._lister) for r in d.get_all_relations() 
                        if r.get_dst_name() == r_name or r.get_src_name() == r_name)
     
-
-    def setup_autocomplete(self):
-        """
-        Sets up autocomplete functionality using the NestedCompleter.
-
-        Returns:
-        NestedCompleter: An instance of NestedCompleter configured with a dictionary representing available commands
-        and their respective subcommands or arguments.
-        """
-        return NestedCompleter.from_nested_dict({
-        'add': {
-            'relation': RelationCompleter(),
-            'class': None,
-            'field': None,
-            'method': None,
-            'param': None
-            
-        },
-        'delete': {
-            'relation': None,
-            'class': None,
-            'field': None,
-            'method': None,
-            'param': None
-            
-        },
-        'rename':{
-            'class': None,
-            'field': None,
-            'method': None,
-            'param': None
-        },
-        'list':{
-            'class': None,
-            'classes': None,
-            'relation': None,
-            'relations': None
-        },
-        'help':{
-            'class': None,
-            'attribute': None,
-            'relation': None
-        },
-        'save': None,
-        'load': None,
-        'undo': None,
-        'redo': None,
-        'quit': None
-        })
